+++ src/network/race_config.hpp
@@ -59,9 +59,9 @@ class RaceVote
     public:
     RaceVote();
 
-    void voteMajor(uint8_t major);
+    void voteMajor(uint32_t major);
     void voteRaceCount(uint8_t count);
-    void voteMinor(uint8_t minor);
+    void voteMinor(uint32_t minor);
     void voteTrack(const std::string &track, uint8_t track_number = 0);
     void voteReversed(bool reversed, uint8_t track_number = 0);
     void voteLaps(uint8_t laps, uint8_t track_number = 0);
@@ -81,8 +81,8 @@ class RaceVote
     uint8_t getLapsVote(uint8_t track_number = 0) const;
 
     private:
-    uint8_t m_major_mode;
-    uint8_t m_minor_mode;
+    uint32_t m_major_mode;
+    uint32_t m_minor_mode;
     uint8_t m_races_count; //!< Stores the number of races that will be in a GP
     bool m_has_voted_major;
     bool m_has_voted_minor;
@@ -97,9 +97,9 @@ class RaceConfig
     RaceConfig();
 
     void setMaxPlayerCount(uint8_t count);
-    void setPlayerMajorVote(uint8_t player_id, uint8_t major);
+    void setPlayerMajorVote(uint8_t player_id, uint32_t major);
     void setPlayerRaceCountVote(uint8_t player_id, uint8_t count);
-    void setPlayerMinorVote(uint8_t player_id, uint8_t minor);
+    void setPlayerMinorVote(uint8_t player_id, uint32_t minor);
     void setPlayerTrackVote(uint8_t player_id, const std::string &track,
                             uint8_t track_nb = 0);
     void setPlayerReversedVote(uint8_t player_id, bool reversed,
